Try to make PR int correct

+++ source/targets/minimal_spanning_tree/include/minimal_spanning_tree/LinkCutTree.h
@@ -48,7 +48,7 @@ public:
   int size() const
   {
     int totalSize = sizeof(*this);
-    totalSize += sizeof(Node) * node.capacity();
+    totalSize += int(sizeof(Node) * node.capacity());
     return totalSize;
   }
